requirements.txt
pandas
requests
python-dotenv
flask
gunicorn

